Pseudocode for data system
It will look something like RimWorld's with the deserializer modeled after Newtonsoft.JSON. Basically it would support the same features with base class overriding and referencing class names in XML for things that used to be modeled with function closures in Lua.

Also, this time the defs will be serialized with the character. A problem with the Lua version was needing to copy fields into the object instance. It would simplify things if the def had the same type as the per-instance data.

It probably would have been better to see how other people solve these problems before being stuck with an inferior solution for two years simply because you didn't come up with anything better the whole time.