Theme specification
The requirements:

1. Toggleable from in-game menu.
2. Can override specific fields without replacing them all.

The problem is that all defs must be loaded at once, so you wouldn't be able apply a theme during the initial load, since the theme def itself has not been loaded.

Maybe:
1. Load all defs.
2. Check for active themes, then merge and re-load any affected entries.
3. Populate DefOfEntries classes.

Or maybe treat ThemeDef as special and load only that first before proceeding with the rest of the loading.

When a reload is requested in-game:
1. Determine affected entries. Take the original XmlNode for the affected entries and create a new Def from the merged result (requires keeping all def XML in memory at all times).
2. Somehow modify the references already active inside DefOfEntries classes/everywhere else, meaning: do an in-place update of the affected entries in DefStore.
3. Use an AssetManager class or similar instead of constructing AssetDrawable instances by hand to support hot reloading. The manager would be able to notify all the drawable assets that they need to be reloaded.

Since mods should also be able to modify defs without the use of C# code or themes, the "merging" mechanism would need to be uniform with the theme override system being proposed.