Refs for Defs?
I want to hotload Defs at runtime.

I'm thinking I will not use this Ref strategy, because it makes the types harder to manage. You'd have to wrap everything that can be hotloaded in a Ref<T> and access the original copy by calling Ref<T>.Value every time.

The problem is well-defined: Only the public/private fields (not properties) should be copied between the defs. If properties were to be involved, all sorts of weird things could happen. Properties make serialization harder, so fields only were already used for def deserialization anyway.

As before, this is how I'm thinking this could be accomplished:

1. Keep all the raw XML in memory, to regenerate the final defs from. Associate each def node with its originating file.
2. When a reload is requested, go to the defset files the defs were sourced from and load them from disk again. Match the corresponding IDs to  the ones in the store.
3. Take the XML nodes corresponding to each overridden def in the theme overrides and merge them together with the original XML nodes somehow. This "merging" feature is important to get right.
4. Generate a new def instance from each merged XML node by concatting all the merged nodes into a "virtual" defset. Since each merged def is a completely new object, we don't have to worry about things like copying the entries of a Dictionary<K, V> that might arise if we were trying to deep copy things instead.
5. Find the original def in the store and copy all the public and private field values from the new instance to the original instance.
6. Increment a count on the original def indicating that it's been hotloaded.
7. Invalidate anything on the old def like cached LOVE fonts (could be done with an OnHotload() method).
8. Notify AssetDrawables that they need to be recreated. This means using an AssetManager or something to create weak references that encapsulate AssetDefs, storing all the references in the manager globally, and then sending the notification for only the AssetDrawables that need updating.

At that point, the theme system should be feature-compatible with the current iteration of ON.