I'm slowly starting to understand the point of the code I removed earlier.
Before, I assumed `TransformComponent` was concerned with things like
visual effects that are parented around an entity. Since Elona doesn't
have those, I simplified it to only hold map coordinates.

But actually, Elona *does* have such parenting. The parenting is
implicit, though:

- Items in containers.
- Ridden allies moving with their riders.
- Tag Team partners.
- Entities in maps.

Months ago, when I was dumbly working on ON/LÖVE still, I was wondering how
I would implement vehicles with more than one character inside. All the
characters would need to move as the vehicle moves. There was no
standard way of making this happen in ON/LÖVE, since parenting occurred
on a container level instead of an entity level. HSP Elona simply updates
a ridden ally's position manually at the end of the rider's movement
action.

Robust has parenting at an entity level. Now I understand the point of
`EntityCoordinates`. This also solves the problem of determining an
entity's map, because maps have associated entities, so whatever parent
entity has a `MapComponent` counts as the map. In effect, this system
represents a set of entities as a scene graph. This is simultaneously a
"what the fuck was I thinking for four years" moment and a "I probably
never would have come up with that independently" moment. It's
dreadfully sad, and internalizing these thoughts has almost made me want
to give up on ON entirely in recent days, but the fact is that I've
learned better now.

It's sad that to me that it feels like I'm no better at architecting
complex systems four years into this project of mine, especially since
I've already spent hundreds of hours internally debating how to solve
those exact same problems - by myself. Everyone says you miss 100% of
the shots you don't take. I went for it, and I chose not to stop
working, and now I feel like an in-between case. I've found that there
is only so much I can come up with on my own. There is no rule that says
that all hard work has to pay off in the end. Maybe the only thing I'm
actually good at is maintaining dogged persistence and not knowing when
to give up. Two years ago, much of the external code that I'm using
didn't exist. Was the serendipity of finding other people's better
solutions the only thing allowing me to succeed? I'd hate to think that
way.

If ON is to be a community-driven project, then I can no longer measure
my means of success by how close the engine is to replicating vanilla
like I did in the past, to give myself a self-assuring pat on the back.
That is the one thing that stresses me out the most. The second most
stressful thing is the fact that I'm not actually good with interacting
with people. This is simply an honest fact. Four years of holing myself
up in my room absorbed in nefia fantasies didn't exactly help that.

My mentality is that code is a means to an end, and I don't mind using
as much of any other person's code under a compatible license if it
accomplishes my objective. End result of this mentality: 258,608 lines
of Lua that accomplished my objective. It was a Pyrrhic victory. This is
going to be a dangerous mindset for something of Elona's scale, so of
course I'll have to impose quality standards similar to Robust's (no
static classes, etc). I also have no intention of ignoring tests. At the
same time, I'm not actually a game designer. I'm already reusing Elona's
original code to recreate parts of the engine. That's just what happens
to interest me, as opposed to writing original games.  I still believe
there is something noble about that. I'm just wondering if I can get to
the programming parts that sustained my interest for several years in
ON/LÖVE before I lose patience again... like now.

The design of all this also serves as a reminder that stacking and
containers are **irreducible** problems. The difference is that the
container system in ON/LÖVE was left in a poorly thought-out state in
favor of just porting more vanilla things to claim that "it works", and
the addition of choosing Lua changed the tractability of the issue from
"we can improve on this later when the need is more pressing" to "it
really was hopeless from the start."

I am now extremely glad I chose not to release ON/LÖVE.

At least when the ridiculously hard parts of stacking, containers, map
management and map prototyping are over with, *then* the shining work
that holds me together can finally break ground.