new mod for dumping weird ideas: LecchoTorte
I like consuming OpenNefia as well as building its foundation. To that end, I wanted a mod that's walled off from content (which is the "official" base Elona distribution) where I can just do whatever the fuck I want in a spontaneous fashion without worrying about running into organization issues like last time.

If the mod is full-featured, then its code shouldn't be a part of LecchoTorte, but should be moved to a new project. (An example would be a mod implementing the entire autopickup feature.) LT is essentially just an incubation ground to try to break things and enjoy oneself.